feat(frontend): RORY REFRESH-PATTERN PERSISTENT LOGIN
This is a big commit that includes a bunch to tidying up, deleting
of loads of console logs/commented out code, but mainly the implementation
of the persistent login of users via the refresh controller.

This returns the userId as well as new tokens, so that the user hook
can then use that Id to fetch the corresponding user to be stored in
local state and allow access to protected areas of the app.